atomic design fixing pattern
