--- front/src/atomes/logo/Logo.css
@@ -1,3 +1,0 @@
-#logo {
-  margin-bottom: 20%;
-}
