+++ front/src/commun/logo/Logo.css
@@ -1,0 +1,3 @@
+#logo {
+  margin-bottom: 20%;
+}
