+++ front/src/game/Game.css
@@ -1,0 +1,3 @@
+.paper {
+  position: absolute;
+}
