+++ front/src/game/board/boxes/cheeseBox/CheeseBox.jsx
