+++ front/src/game/board/boxes/normalBox/NormalBox.jsx
