+++ front/src/game/ranking/ranking.css
