+++ front/src/game/social/social.css
