--- front/src/organismes/game/Game.css
@@ -1,3 +1,0 @@
-.paper {
-  position: absolute;
-}
