--- front/src/organismes/game/board/boxes/centerBox/CenterBox.jsx
@@ -1,18 +1,0 @@
-import React, { Component } from "react";
-
-export default class CenterBox extends Component {
-    constructor (props){
-        super();
-		this.state = {...props.centerObjet}
-        };
-
-    
-    render() {
-      return (
-        <div> CenterBox
-          {this.state.id}
-          <img src={require('./../../../../../commun/images/logoWP.png')}  width = "100" height = "70" />
-          </div>
-          );
-        }
-    }
